feat: finalize PaymentScreen refactor and enhancements
- Split `BasketSummary` and `PaymentForm` into the local `components/` directory for modularity.
- Extracted `parseErrorMessage` utility into a separate function for reusability and clarity.
- Improved readability and maintainability of the `PaymentScreen` logic.
- Prepared the screen for better testability with enhanced structure and utility abstraction.
