+++ src/screens/PaymentScreen/components/PaymentForm/PaymentForm.style.js
@@ -2,9 +2,7 @@ import { StyleSheet } from 'react-native';
 import { spacing } from '@constants/theme';
 
 export default StyleSheet.create({
-  bottomContainer: {
-    flex: 1,
-    justifyContent: 'flex-end',
-    paddingBottom: spacing.lg,
+  formContainer: {
+    marginBottom: spacing.md,
   },
 });
