+++ src/screens/PaymentScreen/components/index.js
@@ -1,0 +1,2 @@
+export { default as BasketSummary } from './BasketSummary';
+export { default as PaymentForm } from './PaymentForm';
